Membuat Class Spouse
Memisahkan atribute Spouse kedalam Class Spouse
