+++ TaxCalculator/src/lib/Employee.java
@@ -5,7 +5,7 @@ import java.time.Month;
 import java.util.LinkedList;
 import java.util.List;
 
-public class Employee extends Salary{
+public class Employee extends Salary, Spouse {
 
 	private enum Gender {
 		LAKI_LAKI,
@@ -22,28 +22,23 @@ public class Employee extends Salary{
 	private boolean isForeigner;
     private Gender gender; 
 
-	private String spouseName;
-	private String spouseIdNumber;
-
 	private List<String> childNames;
 	private List<String> childIdNumbers;
 
-	public Employee(String employeeId, String fullName, Address address, LocalDate joinDate, boolean isForeigner, 
-		Gender gender) {
+	public Employee(String employeeId, String fullName, Address address, Gender gender) {
 		this.employeeId = employeeId;
 		this.fullName = fullName;
 		this.address = address;
-		this.joinDate = joinDate;
-		this.isForeigner = isForeigner;
+		this.isForeigner = address.getIsForeigner();
 		this.gender = gender;
 
 		childNames = new LinkedList<String>();
 		childIdNumbers = new LinkedList<String>();
+		setJoinDate(LocalDate.now());
 	}
 
-	public void setSpouse(String spouseName, String spouseIdNumber) {
-		this.spouseName = spouseName;
-		this.spouseIdNumber = idNumber;
+	public void setJoinDate(LocalDate joinDate){
+		this.joinDate = joinDate;
 	}
 
 	public void addChild(String childName, String childIdNumber) {
